+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/controller/InventoryBatchController.java
@@ -24,7 +24,7 @@ public class InventoryBatchController {
 
     @GetMapping
     public ResponseEntity<ApiResponse<List<InventoryBatchDto>>> getAll(
-            @RequestParam(required = false) Long itemId,
+            @RequestParam(required = false) Integer itemId,
             @RequestParam(required = false) String category) {
         try {
             List<InventoryBatchDto> data;
@@ -43,7 +43,7 @@ public class InventoryBatchController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ApiResponse<InventoryBatchDto>> getById(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<InventoryBatchDto>> getById(@PathVariable Integer id) {
         return inventoryBatchService.findById(id)
                 .map(data -> ResponseEntity.ok(ApiResponse.success(data)))
                 .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND)
@@ -62,7 +62,7 @@ public class InventoryBatchController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ApiResponse<InventoryBatchDto>> update(@PathVariable Long id, @RequestBody InventoryBatchDto dto) {
+    public ResponseEntity<ApiResponse<InventoryBatchDto>> update(@PathVariable Integer id, @RequestBody InventoryBatchDto dto) {
         try {
             return inventoryBatchService.update(id, dto)
                     .map(data -> ResponseEntity.ok(ApiResponse.success(data, "InventoryBatch updated successfully")))
@@ -75,7 +75,7 @@ public class InventoryBatchController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<ApiResponse<Void>> delete(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<Void>> delete(@PathVariable Integer id) {
         try {
             if (inventoryBatchService.deleteById(id)) {
                 return ResponseEntity.ok(ApiResponse.success(null, "InventoryBatch deleted successfully"));
