+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/service/TaskTemplateService.java
@@ -29,11 +29,11 @@ public class TaskTemplateService {
         return taskTemplateRepository.findAll().stream().map(taskTemplateMapper::toDto).collect(Collectors.toList());
     }
 
-    public List<TaskTemplateDto> findByStageId(Long stageId) {
+    public List<TaskTemplateDto> findByStageId(Integer stageId) {
         return taskTemplateRepository.findByStageStageId(stageId).stream().map(taskTemplateMapper::toDto).collect(Collectors.toList());
     }
 
-    public Optional<TaskTemplateDto> findById(Long id) {
+    public Optional<TaskTemplateDto> findById(Integer id) {
         return taskTemplateRepository.findById(id).map(taskTemplateMapper::toDto);
     }
 
@@ -42,14 +42,14 @@ public class TaskTemplateService {
         return taskTemplateMapper.toDto(taskTemplateRepository.save(entity));
     }
 
-    public Optional<TaskTemplateDto> update(Long id, TaskTemplateDto dto) {
+    public Optional<TaskTemplateDto> update(Integer id, TaskTemplateDto dto) {
         if (!taskTemplateRepository.existsById(id)) return Optional.empty();
         TaskTemplate entity = taskTemplateMapper.toEntity(dto);
         entity.setTaskTmpId(id);
         return Optional.of(taskTemplateMapper.toDto(taskTemplateRepository.save(entity)));
     }
 
-    public boolean deleteById(Long id) {
+    public boolean deleteById(Integer id) {
         if (!taskTemplateRepository.existsById(id)) return false;
         taskTemplateRepository.deleteById(id);
         return true;
